requirements.txt
streamlit
langchain
openai
faiss-cpu
pypdf
python-dotenv
tiktoken